+++ packages/components/atoms/f-form-field/stories/Radio.stories.js
@@ -22,7 +22,7 @@ export const RadioComponent = () => ({
             default: text('Value', 'radioLabel')
         },
         isDisabled: {
-            default: select('isDisabled', [null, 'disabled'])
+            default: boolean('isDisabled', false)
         },
         hasError: {
             default: boolean('hasError', false)
@@ -31,21 +31,35 @@ export const RadioComponent = () => ({
             default: boolean('isGrouped', false)
         },
         assistiveText: {
-            default: text('Assistive Text', '')
+            default: text('Assistive Text', 'Change my properties via knobs')
         }
     },
     parameters: {
         notes: 'some documentation here'
     },
-    template:
-        `<form-field
-            :label-text="labelText"
-            :value="value"
-            :has-error="hasError"
-            :is-grouped="isGrouped"
-            input-type="radio"
-            :disabled="isDisabled"
-            :assistiveText="assistiveText"/>`
+    template: `
+        <div>
+            <form-field
+                :assistiveText="assistiveText"
+                :disabled="isDisabled"
+                :has-error="hasError"
+                :is-grouped="isGrouped"
+                :label-text="labelText"
+                :value="value"
+                id="radio-1"
+                input-type="radio"
+                name="group-name" />
+
+                <form-field
+                :disabled="isDisabled"
+                :is-grouped="isGrouped"
+                id="radio-2"
+                input-type="radio"
+                label-text="Second label"
+                name="group-name"
+                value="second-value" />
+        </div>
+        `
 });
 
 RadioComponent.storyName = 'Radio';