+++ packages/components/organisms/f-content-cards/src/components/ContentCards.js
@@ -1,4 +1,4 @@
-import initialiseMetadataDispatcher from '@justeat/f-braze-adapter';
+import BrazeAdapter from '@justeat/f-braze-adapter';
 import { globalisationServices } from '@justeat/f-services';
 import tenantConfigs from '../tenants';
 import {
@@ -97,12 +97,18 @@ export default {
         testId: {
             type: String,
             default: null
+        },
+        tags: {
+            type: String,
+            default: 'content-cards'
         }
     },
     data: () => ({
         cards: [],
         hasLoaded: false,
-        state: STATE_LOADING
+        state: STATE_LOADING,
+        loggingKey: null,
+        defaultLoggingData: {}
     }),
     computed: {
         /**
@@ -132,7 +138,7 @@ export default {
             this.$emit(GET_CARD_COUNT, current.length);
 
             if (current.length && (current.length !== previous.length)) {
-                this.metadataDispatcher.logCardImpressions(this.cards.map(({ id }) => id));
+                this.brazeAdapter.logCardImpressions(this.cards.map(({ id }) => id));
             }
             if ((current.length > 0) && (previous.length === 0)) {
                 this.state = STATE_DEFAULT;
@@ -160,8 +166,15 @@ export default {
             }
         }
     },
-    async mounted () {
-        await this.setupMetadata(this.apiKey, this.userId);
+    mounted () {
+        this.loggingKey = `f-content-cards--${this.userId}--${this.tags}`;
+        this.defaultLoggingData = {
+            tags: this.tags,
+            data: {
+                key: this.loggingKey
+            }
+        };
+        this.setupMetadata(this.apiKey, this.userId);
     },
     /**
      * Sets up dependencies required by descendant components
@@ -205,44 +218,67 @@ export default {
          * @param {String} apiKey
          * @param {String} userId
          * @param {Boolean} enableLogging
-         * @return {Promise<void>}
+         * @return {void}
          **/
         setupMetadata (apiKey, userId, enableLogging = false) {
-            return initialiseMetadataDispatcher({
-                apiKey,
-                userId,
-                enableLogging,
-                enabledCardTypes: [
-                    'Header_Card',
-                    'Promotion_Card_1',
-                    'Promotion_Card_2',
-                    'Recommendation_Card_1',
-                    'Restaurant_FTC_Offer_Card',
-                    'Voucher_Card_1',
-                    'Terms_And_Conditions_Card',
-                    'Terms_And_Conditions_Card_2',
-                    'Anniversary_Card_1',
-                    'Post_Order_Card_1',
-                    'Home_Promotion_Card_1',
-                    'Home_Promotion_Card_2',
-                    'Stamp_Card_1',
-                    'StampCard_Promotion_Card_1'
-                ],
-                brands: this.brands,
-                callbacks: {
-                    handleContentCards: this.metadataContentCards
-                },
-                loggerCallbacks: {
-                    logger: this.handleLogging(this.$logger)
-                }
-            })
-                .then(dispatcher => {
-                    this.metadataDispatcher = dispatcher;
-                })
-                .catch(error => {
-                    this.state = STATE_ERROR;
-                    this.$emit(ON_ERROR, error);
+            try {
+                this.$logger.logInfo(
+                    `Content Cards (setupMetadata) - Attempting to initialise BrazeAdapter. Key: (${this.loggingKey})`,
+                    this.$store,
+                    {
+                        ...this.defaultLoggingData
+                    }
+                );
+
+                this.brazeAdapter = new BrazeAdapter({
+                    apiKey,
+                    userId,
+                    enableLogging,
+                    enabledCardTypes: [
+                        'Header_Card',
+                        'Promotion_Card_1',
+                        'Promotion_Card_2',
+                        'Recommendation_Card_1',
+                        'Restaurant_FTC_Offer_Card',
+                        'Voucher_Card_1',
+                        'Terms_And_Conditions_Card',
+                        'Terms_And_Conditions_Card_2',
+                        'Anniversary_Card_1',
+                        'Post_Order_Card_1',
+                        'Home_Promotion_Card_1',
+                        'Home_Promotion_Card_2',
+                        'Stamp_Card_1',
+                        'StampCard_Promotion_Card_1'
+                    ],
+                    brands: this.brands,
+                    callbacks: {
+                        handleContentCards: this.metadataContentCards
+                    },
+                    logger: this.$logger,
+                    tags: this.tags
                 });
+
+                this.$logger.logInfo(
+                    `Content Cards (setupMetadata) - BrazeAdapter successfully initialised. Key: (${this.loggingKey})`,
+                    this.$store,
+                    {
+                        ...this.defaultLoggingData
+                    }
+                );
+            } catch (e) {
+                this.$logger.logError(
+                    `Content Cards (setupMetadata) - Failed to initialise BrazeAdapter successfully. Key: (${this.loggingKey})`,
+                    this.$store,
+                    {
+                        ...this.defaultLoggingData,
+                        error: {
+                            message: e.message
+                        }
+                    }
+                );
+                this.state = STATE_ERROR;
+                this.$emit(ON_ERROR, e);
+            }
         },
         /**
          * Common method for handling card ingestion to component. Card list length of 0 (after filtering) is considered
@@ -278,6 +314,27 @@ export default {
                 successCallback: () => {
                     this.$emit('on-braze-init', window.appboy); // deprecated -- for backward compatibility
                     this.$emit(ON_METADATA_INIT, window.appboy);
+                    this.$logger.logInfo(
+                        `Content Cards (metadataContentCards) - Successfully received content cards. Key: (${this.loggingKey})`,
+                        this.$store,
+                        {
+                            ...this.defaultLoggingData,
+                            Count: cards.length,
+                            data: {
+                                ...this.defaultLoggingData.data,
+                                source: CARDSOURCE_METADATA
+                            }
+                        }
+                    );
+                },
+                failCallback: () => {
+                    this.$logger.logError(
+                        `Content Cards (metadataContentCards) - Failed to receive content cards array, undefined given. Key: (${this.loggingKey})`,
+                        this.$store,
+                        {
+                            ...this.defaultLoggingData
+                        }
+                    );
                 }
             }, cards);
         },
@@ -287,6 +344,18 @@ export default {
          * @param {Card[]} cards
          **/
         customContentCards (cards) {
+            this.$logger.logInfo(
+                `Content Cards (customContentCards) - Attempting to load cards via custom source. Key: (${this.loggingKey})`,
+                this.$store,
+                {
+                    ...this.defaultLoggingData,
+                    Count: cards.length,
+                    data: {
+                        ...this.defaultLoggingData.data,
+                        source: CARDSOURCE_CUSTOM
+                    }
+                }
+            );
             this.contentCards({
                 source: CARDSOURCE_CUSTOM
             }, cards);
@@ -300,12 +369,23 @@ export default {
             switch (card.source) {
                 case CARDSOURCE_METADATA:
                     this.trackMetadataCardClick(card);
-                    this.metadataDispatcher.logCardClick(card.id);
+                    this.brazeAdapter.logCardClick(card.id);
                     break;
                 case CARDSOURCE_CUSTOM:
                     this.trackCustomCardClick(card);
                     break;
                 default:
+                    this.$logger.logError(
+                        `Content Cards (handleCardClick) - Invalid card source type. Key: (${this.loggingKey})`,
+                        this.$store,
+                        {
+                            ...this.defaultLoggingData,
+                            data: {
+                                ...this.defaultLoggingData.data,
+                                source: CARDSOURCE_METADATA
+                            }
+                        }
+                    );
                     throw new Error('Invalid card source type');
             }
         },
@@ -323,6 +403,17 @@ export default {
                     this.trackCustomCardVisibility(card);
                     break;
                 default:
+                    this.$logger.logError(
+                        `Content Cards (handleCardView) - Invalid card source type. Key: (${this.loggingKey})`,
+                        this.$store,
+                        {
+                            ...this.defaultLoggingData,
+                            data: {
+                                ...this.defaultLoggingData.data,
+                                source: card.source
+                            }
+                        }
+                    );
                     throw new Error('Invalid card source type');
             }
         },
@@ -333,7 +424,7 @@ export default {
          */
         trackMetadataCardClick (card) {
             const event = createMetadataCardEvent('click', card);
-            this.metadataDispatcher.pushShapedEventToDataLayer(this.pushToDataLayer, event);
+            this.brazeAdapter.pushShapedEventToDataLayer(this.pushToDataLayer, event);
         },
 
         /**
@@ -351,7 +442,7 @@ export default {
          */
         trackMetadataCardVisibility (card) {
             const event = createMetadataCardEvent('view', card);
-            this.metadataDispatcher.pushShapedEventToDataLayer(this.pushToDataLayer, event);
+            this.brazeAdapter.pushShapedEventToDataLayer(this.pushToDataLayer, event);
         },
 
         /**
@@ -361,18 +452,6 @@ export default {
         trackCustomCardVisibility (card) {
             const event = createCustomCardEvent('view', card);
             this.pushToDataLayer(event);
-        },
-
-        /**
-         * Handles logging from f-metadata (callback)
-         * @returns {function(*, *=, *=): void}
-         */
-        handleLogging (logger) {
-            return (type, logMessage, payload) => {
-                if (logger && logger[type]) {
-                    logger[type](logMessage, null, payload);
-                }
-            };
         }
     },