+++ packages/components/pages/f-account-info/test-utils/component-objects/f-accountInfo-selectors.js
@@ -1,2 +1,0 @@
-// eslint-disable-next-line import/prefer-default-export
-export const COMPONENT = '[data-test-id="account-info"]';