+++ packages/components/organisms/f-form/CHANGELOG.md
@@ -28,7 +28,7 @@ v0.2.0
 
 v0.1.0
 ------------------------------
-*December 13, 2021*
+*December 21, 2021*
 
 ### Added
 - Generated new component.