+++ packages/components/organisms/f-form/CHANGELOG.md
@@ -22,7 +22,7 @@ v0.2.0
 
 v0.1.0
 ------------------------------
-*December 21, 2021*
+*January 05, 2022*
 
 ### Added
-- Generated new component.
+- Add basic form components.