+++ packages/components/organisms/f-form/CHANGELOG.md
@@ -3,6 +3,21 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+v0.3.0
+------------------------------
+*January 19, 2021*
+
+### Added
+- Add field validations.
+
+
+v0.2.0
+------------------------------
+*January 05, 2022*
+
+### Added
+- Add basic form components.
+
 
 v0.1.0
 ------------------------------